requirements.txt
Flask==3.0.1
google_cloud_aiplatform==1.38.1
langchain==0.1.2
matplotlib==3.8.2
pandas==2.2.0
protobuf==4.25.2
pyspark==3.5.0
seaborn==0.13.1
vertexai==0.0.1
gunicorn
openpyxl
